Call Validator in addition to current validation logic in splitting
The splitting logic has been factored out into its own Validator class,
without touching the code in split.py. This change makes sure that Validator
reports the same errors as the old logic. In a second step, we will remove
the old validation code and just delegate the decision to the Validator class.