Revert "[CI] Build and publish artifacts after tagging a release"
GitHub fails because of a permission problem: The GitHub runner
uses a token that is not actually allowed to create a new release.
Unfortunately, I couldn’t figure out how to create a suitable access
token that won’t expire within a few weeks. We don’t cut releases
that often, so let’s go back to a manual process.

This reverts commit 97e27376a8591a2ef785fa21fa102bb80c44829a.