Add gender to given names
First part of https://github.com/brawer/bern-addresses/issues/277.

Eventually, we’ll use this to infer the gender of people, but this will
need more work. For example, we’ll need to look at titles for `Frl.` or `Witwe`,
and at occupations ending in `-in`; this will provide an additional gender
signal.